backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
PyJWT==2.9.0
sqlparse==0.5.3
django-allauth==0.61.1
django-allauth[google]==0.61.1
django-allauth[socialaccount]==0.61.1
python-dotenv==1.0.1
django-email-verification==0.1.0
